Adding Services & Imps and Repositories changed

+++ logistic-management/logistic-management/src/main/java/com/ingeneo/logisticmanagement/repository/TipoBodegaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ingeneo.logisticmanagement.models.TipoBodega;
 
 @Repository
-public interface TipoBodegaRepository extends JpaRepository<TipoBodega,Long> {
+public interface TipoBodegaRepository extends JpaRepository<TipoBodega,Integer> {
     
 }
