+++ logistic-management/logistic-management/src/main/java/com/ingeneo/logisticmanagement/repository/TipoEnvioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ingeneo.logisticmanagement.models.TipoEnvio;
 
 @Repository
-public interface TipoEnvioRepository  extends JpaRepository<TipoEnvio,Long>{
+public interface TipoEnvioRepository  extends JpaRepository<TipoEnvio,Integer>{
     
 }
